findAllModes algo
This function takes in an array of integers and returns either a number or an array.  If one number shows up with the highest frequency, it returns that.  If all numbers show up the same number of times, it will return [].  If multiple numbers have the same frequency, it will return them in an array